backend/requirements.txt
Django>=5.0,<6.0
django-cors-headers>=4.3
djangorestframework>=3.15
google-generativeai>=0.3.0
python-dotenv>=1.0.0

